+++ src/System.Text.RegularExpressions/ref/System.Text.RegularExpressions.csproj
@@ -4,10 +4,12 @@
   <PropertyGroup>
     <OutputType>Library</OutputType>
     <NuGetTargetMoniker>.NETStandard,Version=v1.7</NuGetTargetMoniker>
-    <DefineConstants Condition="'$(TargetGroup)' == 'netcoreapp1.1'">$(DefineConstants);netcoreapp11</DefineConstants>
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="System.Text.RegularExpressions.cs" />
+  </ItemGroup>
+  <ItemGroup Condition="'$(TargetGroup)' == 'netcoreapp1.1'">
+    <Compile Include="System.Text.RegularExpressions.netcoreapp1.1.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="project.json" />
